+++ packages/expo-doctor/src/checks/DirectoryCheck.ts
@@ -1,6 +1,29 @@
+import chalk from 'chalk';
 import fetch from 'node-fetch';
 
 import { DoctorCheck, DoctorCheckParams, DoctorCheckResult } from './checks.types';
+import { getDirectoryCheckExcludes } from '../utils/doctorConfig';
+
+// Filter out common packages that don't make sense for us to validate on the directory.
+const DEFAULT_PACKAGES_TO_IGNORE = [
+  'react-native',
+  'react',
+  'react-dom',
+  'react-native-web',
+  'jest',
+  /^babel-*/,
+];
+
+export function filterPackages(packages: string[], ignoredPackages: (RegExp | string)[]) {
+  return packages.filter((packageName) => {
+    return ignoredPackages.every((ignoredPackage) => {
+      if (ignoredPackage instanceof RegExp) {
+        return !ignoredPackage.test(packageName);
+      }
+      return ignoredPackage !== packageName;
+    });
+  });
+}
 
 export class DirectoryCheck implements DoctorCheck {
   description = 'Validate packages against React Native Directory package metadata';
@@ -12,9 +35,13 @@ export class DirectoryCheck implements DoctorCheck {
     const newArchUnsupportedPackages: string[] = [];
     const newArchUntestedPackages: string[] = [];
     const unmaintainedPackages: string[] = [];
-
+    const unvalidatedPackages: string[] = [];
     const dependencies = pkg.dependencies ?? {};
-    const packageNames = Object.keys(dependencies);
+    const userDefinedIgnoredPackages = getDirectoryCheckExcludes(pkg);
+    const packageNames = filterPackages(Object.keys(dependencies), [
+      ...DEFAULT_PACKAGES_TO_IGNORE,
+      ...userDefinedIgnoredPackages,
+    ]);
 
     try {
       const response = await fetch('https://reactnative.directory/api/libraries/check', {
@@ -30,6 +57,7 @@ export class DirectoryCheck implements DoctorCheck {
       packageNames.forEach((packageName) => {
         const metadata = packageMetadata[packageName];
         if (!metadata) {
+          unvalidatedPackages.push(packageName);
           return;
         }
 
@@ -55,24 +83,32 @@ export class DirectoryCheck implements DoctorCheck {
 
     if (newArchUnsupportedPackages.length > 0) {
       issues.push(
-        `${newArchUnsupportedPackages.join(', ')} ${newArchUnsupportedPackages.length > 1 ? 'are' : 'is'} not supported on the New Architecture.`
+        `- ${newArchUnsupportedPackages.join(', ')} ${newArchUnsupportedPackages.length > 1 ? 'are' : 'is'} not supported on the New Architecture.`
       );
     }
 
     if (newArchUntestedPackages.length > 0) {
       issues.push(
-        `${newArchUntestedPackages.join(', ')} ${newArchUntestedPackages.length > 1 ? 'are' : 'is'} not tested on the New Architecture.`
+        `- ${newArchUntestedPackages.join(', ')} ${newArchUntestedPackages.length > 1 ? 'are' : 'is'} not tested on the New Architecture.`
       );
     }
 
     if (unmaintainedPackages.length > 0) {
       issues.push(
-        `${unmaintainedPackages.join(', ')} ${unmaintainedPackages.length > 1 ? 'are' : 'is'} unmaintained.`
+        `- ${unmaintainedPackages.join(', ')} ${unmaintainedPackages.length > 1 ? 'are' : 'is'} unmaintained.`
+      );
+    }
+
+    const isSuccessful = issues.length === 0;
+
+    if (unvalidatedPackages.length > 0) {
+      issues.push(
+        `- ${unvalidatedPackages.join(', ')} ${unvalidatedPackages.length > 1 ? 'were' : 'was'} not validated because ${unvalidatedPackages.length > 1 ? 'they are' : 'it is'} not tracked by React Native Directory. You can explicitly skip validating these packages by adding them to ${chalk.bold('expo.doctor.directoryCheck.exclude')} in your package.json.`
       );
     }
 
     return {
-      isSuccessful: !issues.length,
+      isSuccessful,
       issues,
       advice: issues.length
         ? `Use libraries that are actively maintained and support the New Architecture. Find alternatives at https://reactnative.directory`