+++ packages/expo-video/android/src/main/java/expo/modules/video/records/VideoSource.kt
@@ -1,5 +1,6 @@
 package expo.modules.video.records
 import android.content.Context
+import android.net.Uri
 import androidx.annotation.OptIn
 import androidx.media3.common.MediaItem
 import androidx.media3.common.MediaMetadata
@@ -8,12 +9,12 @@ import androidx.media3.exoplayer.source.MediaSource
 import expo.modules.kotlin.records.Field
 import expo.modules.kotlin.records.Record
 import expo.modules.video.UnsupportedDRMTypeException
-import expo.modules.video.utils.buildMediaSourceWithHeaders
+import expo.modules.video.buildMediaSourceWithHeaders
 import java.io.Serializable
 
 @OptIn(UnstableApi::class)
 class VideoSource(
-  @Field var uri: String? = null,
+  @Field var uri: Uri? = null,
   @Field var drm: DRMOptions? = null,
   @Field var metadata: VideoMetadata? = null,
   @Field var headers: Map<String, String>? = null
@@ -37,7 +38,7 @@ class VideoSource(
   fun toMediaItem() = MediaItem
     .Builder()
     .apply {
-      setUri(uri ?: "")
+      setUri(uri)
       setMediaId(toMediaId())
       drm?.let {
         if (it.type.isSupported()) {